refactor: Split code in Contracts to helper and component
No code change except adding propTypes on SecondaryText comp
